+++ tests/pytests/unit/crypt/test_crypt.py
@@ -18,6 +18,36 @@ from tests.support.helpers import dedent
 from . import PRIV_KEY, PRIV_KEY2, PUB_KEY, PUB_KEY2
 
 
+@pytest.fixture
+def key_data():
+    return [
+        "-----BEGIN PUBLIC KEY-----",
+        "MIIBIjANBgkqhkiG9w0BAQEFAAOCAQ8AMIIBCgKCAQEAoe5QSDYRWKyknbVyRrIj",
+        "rm1ht5HgKzAVUber0x54+b/UgxTd1cqI6I+eDlx53LqZSH3G8Rd5cUh8LHoGedSa",
+        "E62vEiLAjgXa+RdgcGiQpYS8+Z2RvQJ8oIcZgO+2AzgBRHboNWHTYRRmJXCd3dKs",
+        "9tcwK6wxChR06HzGqaOTixAuQlegWbOTU+X4dXIbW7AnuQBt9MCib7SxHlscrqcS",
+        "cBrRvq51YP6cxPm/rZJdBqZhVrlghBvIpa45NApP5PherGi4AbEGYte4l+gC+fOA",
+        "osEBis1V27djPpIyQS4qk3XAPQg6CYQMDltHqA4Fdo0Nt7SMScxJhfH0r6zmBFAe",
+        "BQIDAQAB",
+        "-----END PUBLIC KEY-----",
+    ]
+
+
+@pytest.mark.parametrize("linesep", ["\r\n", "\r", "\n"])
+def test__clean_key(key_data, linesep):
+    tst_key = linesep.join(key_data)
+    chk_key = "\n".join(key_data)
+    assert salt.crypt.clean_key(tst_key) == salt.crypt.clean_key(chk_key)
+
+
+@pytest.mark.parametrize("linesep", ["\r\n", "\r", "\n"])
+def test__clean_key_mismatch(key_data, linesep):
+    tst_key = linesep.join(key_data)
+    tst_key = tst_key.replace("5", "4")
+    chk_key = "\n".join(key_data)
+    assert salt.crypt.clean_key(tst_key) != salt.crypt.clean_key(chk_key)
+
+
 def test_get_rsa_pub_key_bad_key(tmp_path):
     """
     get_rsa_pub_key raises InvalidKeyError when encoutering a bad key
